move mark all as read to header and remove it from dropdown
